+++ src/Network/Network/Models/PSApplicationGatewayFirewallPolicyExclusionManagedRuleGroup.cs
@@ -13,22 +13,17 @@
 // limitations under the License.
 //
 
-using Microsoft.Azure.Management.Network.Models;
 using Microsoft.WindowsAzure.Commands.Common.Attributes;
-using Newtonsoft.Json;
 using System.Collections.Generic;
 
 namespace Microsoft.Azure.Commands.Network.Models
 {
-    public partial class PSApplicationGatewayFirewallExclusion
+    public partial class PSApplicationGatewayFirewallPolicyExclusionManagedRuleGroup
     {
         [Ps1Xml(Target = ViewControl.Table)]
-        public string MatchVariable { get; set; }
+        public string RuleGroupName { get; set; }
+
         [Ps1Xml(Target = ViewControl.Table)]
-        public string SelectorMatchOperator { get; set; }
-        [Ps1Xml(Target = ViewControl.Table)]
-        public string Selector { get; set; }
-        [Ps1Xml(Target = ViewControl.Table)]
-        public List<PSApplicationGatewayFirewallPolicyExclusionManagedRuleSet> ExclusionManagedRuleSets { get; set; }
+        public List<PSApplicationGatewayFirewallPolicyExclusionManagedRule> Rules { get; set; }
     }
 }